requirements.txt
Django==2.2
djangorestframework==3.11.*
djangorestframework-simplejwt==4.4.*
Pillow==6.2.*
pytz==2019.*
sqlparse==0.3.*
xlwt==1.3.*
python-dateutil==2.8.*
